+++ mne/datasets/__init__.py
@@ -5,3 +5,4 @@ from . import sample
 from . import megsim
 from . import spm_face
 from . import eegbci
+from . import somato